เหมือนตอนกด commit sync ไฟล์ที่ตัวเองทำ แก้มัน sync ไฟล์พวกนี้ลงมาด้วย

+++ protected/modules/learning/controllers/ManageController.php
@@ -47,7 +47,7 @@ class ManageController extends Controller {
     }
 
     public function actionInsertLearningGroup($id = null) {
-        $file = new Upload();
+        $file = new Upload2image();
         if ($id == null) {
             $model = new LearningGroup();
             $model->unsetAttributes();
@@ -61,20 +61,40 @@ class ManageController extends Controller {
             $model->attributes = $_POST['LearningGroup'];
 
             //รูปภาพภาษาไทย
-            $file->image = $_POST['Upload']['image'];
-            $file->image = CUploadedFile::getInstance($file, 'image');
-            if ($file->image != NULL) {
-                $model->pic = $file->image;
+//            $file->image = $_POST['Upload2image']['image'];
+            $file1 = CUploadedFile::getInstance($file, 'image');
+
+            if ($file1 != NULL) {
+                $dir = './file/learning/';
+                if (isset($model->pic)) {// ถ้ามีไฟล์อัพมาใหม่ ต้องลบไฟลเก่าก่อน แล้วค่อยอัพไฟล์ใหม่กลับเข้าไป
+                    if (fopen($dir . $model->pic, 'w')) {
+                        if (unlink($dir . $model->pic)) {
+                            
+                        }
+                    }
+                }
+                $file1name = rand('000', '999') . $file1->name;
+                $model->pic = $file1name;
             } else {
                 if ($model->pic == NULL)
                     $model->pic = 'default.jpg';
             }
             //--------------
             //รูปภาพภาษาอังกฤษ
-            $file->image2 = $_POST['Upload']['image2'];
-            $file->image2 = CUploadedFile::getInstance($file, 'image2');
-            if ($file->image2 != NULL) {
-                $model->pic_en = $file->image2;
+//            $file->image2 = $_POST['Upload2image']['image2'];
+            $file2 = CUploadedFile::getInstance($file, 'image2');
+
+            if ($file2 != NULL) {
+                $dir = './file/learning/';
+                if (isset($model->pic_en)) {// ถ้ามีไฟล์อัพมาใหม่ ต้องลบไฟลเก่าก่อน แล้วค่อยอัพไฟล์ใหม่กลับเข้าไป
+                    if (fopen($dir . $model->pic_en, 'w')) {
+                        if (unlink($dir . $model->pic_en)) {
+                            
+                        }
+                    }
+                }
+                $file2name = rand('000', '999') . $file2->name;
+                $model->pic_en = $file2name;
             } else {
                 if ($model->pic_en == NULL)
                     $model->pic_en = 'default.jpg';
@@ -84,24 +104,24 @@ class ManageController extends Controller {
             $model->validate();
             if ($model->getErrors() == null) {
                 if ($model->save()) {
-                    if ($file->image != NULL) {
-                        $dir = './file/learning/';
-                        if (!file_exists($dir))
-                            mkdir($dir, 0777);
+                    $dir = './file/learning/';
+                    if ($file1 != NULL) {
 
-                        $image = $dir . '/' . $file->image;
+                        $image = $dir . $file1name;
 
-                        $file->image->saveAs($image);
+                        $file1->saveAs($image);
                     }
 
-                    if ($file->image2 != NULL) {
-                        $dir = './file/learning/';
+                    if ($file2 != NULL) {
+
+
+
                         if (!file_exists($dir))
-                            mkdir($dir, 0777);
+                            chmod($dir, 0777);
 
-                        $image = $dir . '/' . $file->image2;
+                        $image = $dir . $file2name;
 
-                        $file->image2->saveAs($image);
+                        $file2->saveAs($image);
                     }
                     echo "
                         <script>
@@ -125,29 +145,30 @@ class ManageController extends Controller {
 
     public function actionDelLearningGroup($id = null) {
         $model = LearningGroup::model()->findByPk($id);
-
-        if ($model->pic != 'default.jpg') {
-//            $file_paht = './file/learning';
-//            if (!file_exists($file_paht))
-//                mkdir($file_paht, 0777);
-//
-//            unlink($file_paht . '/' . $model->pic);
-            $file_paht = './file/learning/' . $model->pic;
-            if (fopen($file_paht, 'w'))
-                unlink($file_paht);
-        }
-        if ($model->pic != $model->pic_en && $model->pic_en != null && $model->pic != null) {
-            if ($model->pic_en != 'default.jpg') {
-                $file_paht = './file/learning/' . $model->pic_en;
-//                if (!file_exists($file_paht))
-//                    mkdir($file_paht, 0777);
+        $count = Learning::model()->count('group_id=:group_id', array(':group_id' => $model->id));
+        if (empty($count)) {
+            if ($model->pic != 'default.jpg') {
+                $file_paht = './file/learning/' . $model->pic;
+                chmod($file_paht, 0777);
                 if (fopen($file_paht, 'w'))
                     unlink($file_paht);
             }
-        }
+            if ($model->pic != $model->pic_en && $model->pic_en != null && $model->pic != null) {
+                if ($model->pic_en != 'default.jpg') {
+                    $file_paht = './file/learning/' . $model->pic_en;
+                    chmod($file_paht, 0777);
+//                if (!file_exists($file_paht))
+//                    mkdir($file_paht, 0777);
+                    if (fopen($file_paht, 'w'))
+                        unlink($file_paht);
+                }
+            }
 
-        if ($model->delete())
-            echo "ลบข้อมูลเรียบร้อย";
+            if ($model->delete())
+                echo "ลบข้อมูลเรียบร้อย";
+        }else {
+            echo "ไม่สามารถลบข้อมูลได้ มีข้อมูลอ้างอิงอยู่";
+        }
     }
 
     public function actionLearning() {
@@ -164,8 +185,8 @@ class ManageController extends Controller {
     }
 
     public function actionInsertLearning($id = null) {
-        $upload = new Upload();
-        $upload->unsetAttributes();
+        $upload = new UploadPDF();
+//        $upload->unsetAttributes();
 
         if ($id == null) {
             $model = new Learning();
@@ -178,15 +199,17 @@ class ManageController extends Controller {
             $modelFile->unsetAttributes();
 
             $messageAlert = 'เพิ่มข้อมูลเรียบร้อย';
+            $link = '/learning/manage/inserLearning';
         } else {
             $model = Learning::model()->findByPk($id);
             $modelVideo = LearningVideo::model()->find('main_id = ' . $id);
             $messageAlert = 'แก้ไขข้อมูลเรียบร้อย';
+            $link = '/learning/default/lesson/id/' . $model->id;
 
             $modelFile = LearningFile::model()->find('main_id = ' . $id);
         }
 
-        if (isset($_POST['Learning']) && isset($_POST['LearningVideo'])) {
+        if (isset($_POST['Learning']) && isset($_POST['LearningVideo']) && isset($_POST['UploadPDF'])) {
             $model->attributes = $_POST['Learning'];
             $model->author = Yii::app()->user->id;
             $model->guide_status = '0';
@@ -196,29 +219,29 @@ class ManageController extends Controller {
             $modelVideo->attributes = $_POST['LearningVideo'];
             $modelVideo->video = str_replace('watch?v=', 'embed/', $modelVideo->video);
 
-            $upload->attributes = $_POST['Upload'];
+            $upload->attributes = $_POST['UploadPDF'];
+            
             if (isset($id) && $upload->file == null) { // ถ้าแก้ไขบทเรียนเดิม ตั้งค่า default ของ Model Upload = ค่าที่ต้องการแก้ไข
                 $upload->file = $modelFile->path;
             }
 
+            $upload->file = CUploadedFile::getInstance($upload, 'file');
             $model->validate();
             $modelVideo->validate();
-            $upload->validate();
-
-            if ($upload->getErrors() == null) {
-                $file = CUploadedFile::getInstance($upload, 'file');
-            }
+//            $upload->validate();
 
             if ($model->getErrors() == null && $modelVideo->getErrors() == null && $upload->getErrors() == null) {
+                $file = CUploadedFile::getInstance($upload, 'file');
                 if ($model->save()) {
-
                     $modelVideo->main_id = $model->id;
                     if ($modelVideo->save()) {
                         if (isset($file)) {
                             $path = './file/learning/pdf/';
-                            if (isset($id)) {// ถ้ามีไฟล์อัพมาใหม่ ต้องลบไฟลเก่าก่อน แล้วค่อยอัพไฟล์ใหม่กลับเข้าไป
-                                if (unlink($path . '/' . $modelFile->path)) {
-                                    $modelFile->delete();
+                            if (isset($modelFile->path)) {// ถ้ามีไฟล์อัพมาใหม่ ต้องลบไฟลเก่าก่อน แล้วค่อยอัพไฟล์ใหม่กลับเข้าไป
+                                if (fopen($path . $modelFile->path, 'w')) {
+                                    if (unlink($path . $modelFile->path)) {
+                                        $modelFile->delete();
+                                    }
                                 }
                             }
 
@@ -227,10 +250,12 @@ class ManageController extends Controller {
                                     mkdir($path, 0777, true);
                                     chmod($path, 0777);
                                 }
-                                if ($file->saveAs($path . '/' . $file->name)) {
+                                $filename = rand('000', '999') . $file->name;
+
+                                if ($file->saveAs($path . '/' . $filename)) {
                                     $modelFile = new LearningFile();
                                     $modelFile->main_id = $model->id;
-                                    $modelFile->path = $file->name;
+                                    $modelFile->path = $filename;
 
                                     $modelFile->save();
                                 }
@@ -239,7 +264,7 @@ class ManageController extends Controller {
                         echo "
                         <script>
                         alert('" . Yii::t('language', $messageAlert) . "');
-                        window.location='/learning/manage/InsertLearning';
+                        window.location='$link';
                         </script>
                         ";
                     }
@@ -253,6 +278,10 @@ class ManageController extends Controller {
 //                print_r(array($model->getErrors(), $modelVideo->getErrors()));
 //                echo "</pre>";
             }
+        } else {
+//            echo "<pre>";
+//            print_r(array($model->getErrors(),$upload->getErrors()));
+//            echo "</pre>";
         }
 
         $this->render('_insert_learning', array(
@@ -269,6 +298,7 @@ class ManageController extends Controller {
 
         if ($modelFile->file != 'default.jpg') {
             $file_paht = './file/learning/pdf/' . $modelFile->file;
+            mkdir($file_paht, 0777);
             if (fopen($file_paht, 'w'))
                 unlink($file_paht);
         }
@@ -278,4 +308,5 @@ class ManageController extends Controller {
     }
 
 }
+
 ?>
